orderss

+++ teknofest/Models/CartModel.cs
@@ -5,7 +5,7 @@
         public int CartId { get; set; }
         public List<CartItemModel> CartItems { get; set; }
 
-        public decimal TotalPrice()
+        public int TotalPrice()
         {
             return CartItems.Sum(i => i.Price * i.Quantity);
         }
@@ -16,7 +16,7 @@
         public int CartItemId { get; set; }
         public int FoodId { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
         public string ImageUrl { get; set; }
         public int Quantity { get; set; }
     }
